requirements.txt
numpy==1.25.2
pyro_ppl==1.8.4+b8e0c3ce
scikit_learn==1.2.1
scipy==1.11.1
torch==1.13.0
tqdm==4.64.1
